# server/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Blockchain-based certificate issuance and verification system",
        "title": "Certificate System API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/certificates/batch-anchor": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Anchor multiple certificates to blockchain using Merkle tree and optionally send emails to students",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Certificates"
                ],
                "summary": "Anchor certificates to blockchain",
                "parameters": [
                    {
                        "description": "Batch anchor request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.BatchAnchorRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.BatchAnchorResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/certificates/bulk-download": {
            "get": {
                "description": "Download all certificates for a university as a ZIP file",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/zip"
                ],
                "tags": [
                    "Certificates"
                ],
                "summary": "Bulk download certificates",
                "parameters": [
                    {
                        "type": "string",
                        "description": "University ID",
                        "name": "university_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Password",
                        "name": "password",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/certificates/{certID}/download": {
            "get": {
                "description": "Download the PDF certificate from IPFS",
                "produces": [
                    "application/pdf"
                ],
                "tags": [
                    "certificates"
                ],
                "summary": "Download certificate PDF",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Certificate ID",
                        "name": "certID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    }
                }
            }
        },
        "/certificates/{id}/download": {
            "get": {
                "description": "Download a single certificate PDF by ID",
                "produces": [
                    "application/pdf"
                ],
                "tags": [
                    "Certificates"
                ],
                "summary": "Download certificate PDF",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Certificate ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handlers.BatchAnchorRequest": {
            "type": "object",
            "required": [
                "password",
                "university_id"
            ],
            "properties": {
                "max_certs": {
                    "description": "Optional, default 1000",
                    "type": "integer"
                },
                "password": {
                    "type": "string"
                },
                "send_emails": {
                    "type": "boolean"
                },
                "university_id": {
                    "type": "string"
                }
            }
        },
        "handlers.BatchAnchorResponse": {
            "type": "object",
            "properties": {
                "cardano_txid": {
                    "type": "string",
                    "example": "f75eac8ac7b3..."
                },
                "certificates": {
                    "type": "integer",
                    "example": 10
                },
                "cost": {
                    "type": "string",
                    "example": "0.17 ADA (~$0.08)"
                },
                "cost_per_cert": {
                    "type": "string",
                    "example": "$0.008000"
                },
                "email_errors": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "email_warning": {
                    "type": "string",
                    "example": "2 emails failed to send"
                },
                "emails_sent": {
                    "type": "integer",
                    "example": 10
                },
                "explorer_url": {
                    "type": "string",
                    "example": "https://preview.cardanoscan.io/transaction/..."
                },
                "merkle_root": {
                    "type": "string",
                    "example": "abc123def456..."
                },
                "message": {
                    "type": "string",
                    "example": "✅ 10 certificates anchored and 10 emails sent!"
                },
                "network": {
                    "type": "string",
                    "example": "preview"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                },
                "wallet": {
                    "type": "string",
                    "example": "addr_test1..."
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}